+++ controllers/views/error.js
@@ -6,7 +6,7 @@ const { _M_ } = require('../middleware/plugins');
 // gateway
 function errorHandler (err, req, res, next) {
   if (['dev', 'test'].indexOf(process.env['NODE_ENV']) !== -1) console.log(err.stack);
-  if (_U_.error.hasOwnProperty(err.name) && terminal[err.name]) {
+  if (_U_.error.hasOwnProperty(err.name) && !!terminal[err.name]) {
     return terminal[err.name](err, req, res, next);
   } else return res.render('./theme/error', { err });
 }
@@ -40,7 +40,7 @@ terminal.MongoError = (err, req, res, next) => {
 
 terminal.HttpError = (err, req, res, next) => {
   return _M_.doNotCrawled(req, res, () => {
-    return res.status(err.status).render('./theme/error', {err});
+    return res.status(err.code).render('./theme/error', {err});
   });
 };